Rename types. Code review
