+++ src/components/offers-list/components/offers-item/offers-item.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import {AppRoute} from '../../const';
+import {AppRoute} from '../../../../const';
 
 const OffersItem = () => {
   const offerRoute = `${AppRoute.OFFER}`;
